backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
celery[redis]==5.3.6
httpx==0.27.0
chromadb==0.5.0
sentence-transformers==2.7.0
langchain-text-splitters==0.2.0
python-docx==1.1.0
reportlab==4.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pyvis==0.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
beautifulsoup4==4.12.3
python-multipart==0.0.9
chardet==5.2.0
markdown-it-py==3.0.0
redis==5.0.0
